+++ src/main/java/com/talhacgdem/twitterclone/security/JwtUserDetails.java
@@ -15,12 +15,12 @@ import java.util.List;
 @Setter
 public class JwtUserDetails implements UserDetails {
 
-	public Integer id;
+	public Long id;
 	private String username;
 	private String password;
 	private Collection<? extends GrantedAuthority> authorities;
 	
-    private JwtUserDetails(Integer id, String username, String password, Collection<? extends GrantedAuthority> authorities) {
+    private JwtUserDetails(Long id, String username, String password, Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.username = username;
         this.password = password;
